requirements.txt
setuptools~=73.0.1
wheel~=0.44.0
twine~=5.1.1
mypy~=1.11.1
black~=24.8.0
tox~=4.18.0
versioneer~=0.29
pytest~=8.3.2
pyinstaller~=6.10.0
pyyaml~=6.0.2
types-PyYAML~=6.0.12.20240808
jsonschema~=4.20.0

template-project-utils~=0.1.0